requirements.txt
pyserial
pandas
matplotlib
click
